move dead code elimination code to MastForest
