+++ core/src/mast/mod.rs
@@ -1,5 +1,8 @@
-use alloc::vec::Vec;
-use core::{fmt, ops::Index};
+use alloc::{
+    collections::{BTreeMap, BTreeSet, VecDeque},
+    vec::Vec,
+};
+use core::{fmt, mem, ops::Index};
 
 use miden_crypto::hash::rpo::RpoDigest;
 
@@ -136,9 +139,172 @@ impl MastForest {
             self.roots.push(new_root_id);
         }
     }
+
+    /// Removes all nodes that are unreachable from all procedures are removed.
+    ///
+    /// It also returns the map from old node IDs to new node IDs; or `None` if the `MastForest` was
+    /// unchanged. Any [`MastNodeId`] used in reference to the old [`MastForest`] should be remapped
+    /// using this map.
+    pub fn prune_unreachable_nodes(&mut self) -> Option<BTreeMap<MastNodeId, MastNodeId>> {
+        let live_ids = self.compute_live_ids();
+        if live_ids.len() == self.num_nodes() as usize {
+            return None;
+        }
+
+        let old_nodes = mem::take(&mut self.nodes);
+        let old_root_ids = mem::take(&mut self.roots);
+        let (live_nodes, id_remappings) = prune_nodes(old_nodes, live_ids);
+
+        self.add_remapped_live_nodes(live_nodes, &id_remappings);
+        self.add_remapped_roots(old_root_ids, &id_remappings);
+        Some(id_remappings)
+    }
+}
+
+/// Helpers
+impl MastForest {
+    /// Computes the set of [`MastNodeId`]s which are "live"; that is, that are reached by at least
+    /// one procedure in the MAST forest.
+    fn compute_live_ids(&self) -> BTreeSet<MastNodeId> {
+        let mut live_ids = BTreeSet::new();
+
+        let mut worklist = VecDeque::from_iter(self.procedure_roots().iter().copied());
+        while let Some(mast_node_id) = worklist.pop_front() {
+            if !live_ids.insert(mast_node_id) {
+                continue;
+            }
+
+            match &self[mast_node_id] {
+                MastNode::Join(node) => {
+                    worklist.push_back(node.first());
+                    worklist.push_back(node.second());
+                },
+                MastNode::Split(node) => {
+                    worklist.push_back(node.on_true());
+                    worklist.push_back(node.on_false());
+                },
+                MastNode::Loop(node) => {
+                    worklist.push_back(node.body());
+                },
+                MastNode::Call(node) => {
+                    worklist.push_back(node.callee());
+                },
+                MastNode::Block(_) | MastNode::Dyn | MastNode::External(_) => (),
+            }
+        }
+
+        live_ids
+    }
+
+    /// Adds all live nodes to the internal set of nodes, remapping all [`MastNodeId`] references in
+    /// those nodes.
+    ///
+    /// # Panics
+    /// - Panics if the internal set of nodes is not empty.
+    fn add_remapped_live_nodes(
+        &mut self,
+        live_nodes: Vec<MastNode>,
+        id_remappings: &BTreeMap<MastNodeId, MastNodeId>,
+    ) {
+        assert!(self.nodes.is_empty());
+
+        // Add each live node to the new MAST forest, making sure to rewrite any outdated internal
+        // `MastNodeId`s
+        for live_node in live_nodes {
+            match &live_node {
+                MastNode::Join(join_node) => {
+                    let first_child =
+                        id_remappings.get(&join_node.first()).copied().unwrap_or(join_node.first());
+                    let second_child = id_remappings
+                        .get(&join_node.second())
+                        .copied()
+                        .unwrap_or(join_node.second());
+
+                    self.add_join(first_child, second_child).unwrap();
+                },
+                MastNode::Split(split_node) => {
+                    let on_true_child = id_remappings
+                        .get(&split_node.on_true())
+                        .copied()
+                        .unwrap_or(split_node.on_true());
+                    let on_false_child = id_remappings
+                        .get(&split_node.on_false())
+                        .copied()
+                        .unwrap_or(split_node.on_false());
+
+                    self.add_split(on_true_child, on_false_child).unwrap();
+                },
+                MastNode::Loop(loop_node) => {
+                    let body_id =
+                        id_remappings.get(&loop_node.body()).copied().unwrap_or(loop_node.body());
+
+                    self.add_loop(body_id).unwrap();
+                },
+                MastNode::Call(call_node) => {
+                    let callee_id = id_remappings
+                        .get(&call_node.callee())
+                        .copied()
+                        .unwrap_or(call_node.callee());
+
+                    if call_node.is_syscall() {
+                        self.add_syscall(callee_id).unwrap();
+                    } else {
+                        self.add_call(callee_id).unwrap();
+                    }
+                },
+                MastNode::Block(_) | MastNode::Dyn | MastNode::External(_) => {
+                    self.add_node(live_node).unwrap();
+                },
+            }
+        }
+    }
+
+    /// Remaps and adds all old root ids to the internal set of roots.
+    ///
+    /// # Panics
+    /// - Panics if the internal set of roots is not empty.
+    fn add_remapped_roots(
+        &mut self,
+        old_root_ids: Vec<MastNodeId>,
+        id_remappings: &BTreeMap<MastNodeId, MastNodeId>,
+    ) {
+        assert!(self.roots.is_empty());
+
+        for old_root_id in old_root_ids {
+            let new_root_id = id_remappings.get(&old_root_id).copied().unwrap_or(old_root_id);
+            self.make_root(new_root_id);
+        }
+    }
+}
+
+/// Returns the set of nodes that are live, as well as the mapping from "old ID" to "new ID" for all
+/// live nodes.
+fn prune_nodes(
+    mast_nodes: Vec<MastNode>,
+    live_ids: BTreeSet<MastNodeId>,
+) -> (Vec<MastNode>, BTreeMap<MastNodeId, MastNodeId>) {
+    // Note: this allows us to safely use `usize as u32`, guaranteeing that it won't wrap around.
+    assert!(mast_nodes.len() < u32::MAX as usize);
+
+    let mut pruned_nodes = Vec::with_capacity(mast_nodes.len());
+    let mut id_remappings = BTreeMap::new();
+
+    for (old_node_index, old_node) in mast_nodes.into_iter().enumerate() {
+        let old_node_id: MastNodeId = (old_node_index as u32).into();
+
+        if live_ids.contains(&old_node_id) {
+            let new_node_id: MastNodeId = (pruned_nodes.len() as u32).into();
+            id_remappings.insert(old_node_id, new_node_id);
+
+            pruned_nodes.push(old_node);
+        }
+    }
+
+    (pruned_nodes, id_remappings)
 }
 
 // ------------------------------------------------------------------------------------------------
+
 /// Public accessors
 impl MastForest {
     /// Returns the [`MastNode`] associated with the provided [`MastNodeId`] if valid, or else
@@ -184,15 +350,6 @@ impl MastForest {
     /// Returns the number of nodes in this MAST forest.
     pub fn num_nodes(&self) -> u32 {
         self.nodes.len() as u32
-    }
-}
-
-/// Destructors
-impl MastForest {
-    pub fn into_parts(self) -> (Vec<MastNode>, Vec<MastNodeId>) {
-        let Self { nodes, roots } = self;
-
-        (nodes, roots)
     }
 }
